update to edition 2021 & set rust-version
so far the crate used the default edition (2015).

upgrading to the newest edition before having the first major release
creates a stable platform on which to build in the future. the current
edition is also already two years old, so it is likely that most
consumers are already using a compatible rust version. for those which
might not have upgraded it remains an option to use an existing release
of the crate for the time being.

setting [`rust-version`] allows having cleaner error messages for those
using older rust compilers - though as of now this will not be
particularly helpful as the field is only supported starting with rust
1.56 which is also the version introducing support for the 2021 edition.
nevertheless, already specifying it now creates a base for future MSRV
updates.

note: all code changes were done by running `cargo fix --edition` as per
the [upgrade guide].

[upgrade guide]: https://doc.rust-lang.org/edition-guide/editions/transitioning-an-existing-project-to-a-new-edition.html
[`rust-version`]: https://doc.rust-lang.org/cargo/reference/manifest.html#the-rust-version-field