Add missing `unsafe` to murmur3::Hasher::push
This commit adds the `unsafe` keyword to the private `murmur3::Hasher::push`
function and associated comments, since the function will result in a buffer
overflow if `self.index.usize() + buf.len() > 4`. This change makes no
difference to the public API or runtime behavior, but it clarifies the safety
requirements of the function.